BackgroundWriter must not keep itself alive.
The TaskScheduler holds a permanent reference to the BackgroundWriter
task. If the block chain is not disposed then the task will never
terminate. Symptom: NCrunch task runners permanently busy running tests
which never complete.

By separating the WritableBlockChain from the Writer task, we can use the
finalisation of the block chain as a trigger for the Writer to shut down.
LivenessIndicator encapsulates the abort token as well as a weak reference
to a marker object, which is strongly referenced by the Writer via a
struct. When the struct releases the strong reference, it is only a matter
of time before the Writer notices that its parent block chain has died and
shuts itself down too.