The background writer should wait on available work.
Spinning is not made significantly less wasteful by Task.Yield(). The
background writer should wait asynchronously on an event instead.

I couldn't find an asynchronous AutoResetEvent implementation so I wrote
one myself. Does such a thing exist? Or perhaps there's a more elegant way
of doing this?

TaskCompletionSources and locks are tricky beasts. It is an extremely
risky thing to complete the former inside the latter and usually results
in a thoroughly confusing deadlock some way down the line. For this reason,
the AwaitableAutoResetEvent never calls into the TCS while inside its lock.