CloseAsync should always throw exceptions asynchronously.
I'm not sure how to test this case properly, although manually inserting
a 'throw new Exception()' proves that adding the relevant keywords does
actually have the intended effect. Suggestions welcome.

It's not immediately obvious that CloseAsync is performing the commits
synchronously. I suspect that 'async' and 'await' work a bit like 'yield',
in that the output format/location of the method is consistent but nothing
clever happens until the 'yield'/'await' is hit. That is to say,
everything's synchronous up until the 'await', but all exceptions will be
reliably dumped in the Task no matter where they came from. Also welcome
are suggestions on how to test this without making a mess of the code.